+++ src/components/SideList/SideList.js
@@ -1,22 +1,22 @@
 import React, { useEffect, useContext } from "react";
 import { Store } from "../../store/index";
-import { fetchRelatedData } from "../../apis/index";
+// import { fetchRelatedData } from "../../apis/index";
 import SideListItem from "../SideListItem/SideListItem";
 import Style from "./SideList.module.scss";
 
 const SideList = () => {
   const { globalState, setGlobalState } = useContext(Store);
-  const setRelatedVideo = async (id) => {
-    await fetchRelatedData(id).then((res) => {
-      setGlobalState({
-        type: "SET_RELATED",
-        payload: { related: res.data.items },
-      });
-    });
-  };
-  useEffect(() => {
-    setRelatedVideo(globalState.selected.id);
-  }, [globalState.selected]);
+  // const setRelatedVideo = async (id) => {
+  //   await fetchRelatedData(id).then((res) => {
+  //     setGlobalState({
+  //       type: "SET_RELATED",
+  //       payload: { related: res.data.items },
+  //     });
+  //   });
+  // };
+  // useEffect(() => {
+  //   setRelatedVideo(globalState.selected.id);
+  // }, [globalState.selected]);
   return (
     <div className={Style.sidenav.id}>
       {globalState.related ? (
