+++ src/components/VideoDetail/VideoDetail.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from "react";
-import { useLocation } from "react-router-dom";
-import { fetchSelectedData } from "../../apis";
+// import { useLocation } from "react-router-dom";
+// import { fetchSelectedData } from "../../apis";
 import { Store } from "../../store/index";
 import VideoPlay from "../VideoPlay/VideoPlay";
 import Style from "./VideoDetail.module.scss";
@@ -8,20 +8,20 @@ import Linkify from "react-linkify";
 
 const VideoDetail = () => {
   const { globalState, setGlobalState } = useContext(Store);
-  const location = useLocation();
-  const setSelectedVideo = async () => {
-    const searchParams = new URLSearchParams(location.search);
-    const id = searchParams.get("v");
-    await fetchSelectedData(id).then((res) => {
-      const item = res.data.items.shift();
-      setGlobalState({ type: "SET_SELECTED", payload: { selected: item } });
-      console.log("res", res);
-    });
-  };
+  // const location = useLocation();
+  // const setSelectedVideo = async () => {
+  //   const searchParams = new URLSearchParams(location.search);
+  //   const id = searchParams.get("v");
+  //   await fetchSelectedData(id).then((res) => {
+  //     const item = res.data.items.shift();
+  //     setGlobalState({ type: "SET_SELECTED", payload: { selected: item } });
+  //     console.log("res", res);
+  //   });
+  // };
 
-  useEffect(() => {
-    setSelectedVideo();
-  }, [location.search]);
+  // useEffect(() => {
+  //   setSelectedVideo();
+  // }, [location.search]);
   return globalState.selected && globalState.selected.id ? (
     <div className={Style.wrap}>
       <VideoPlay id={globalState.selected.id} />
